@@ -155,36 +163,38 @@
     "css-minimizer-webpack-plugin": "^7.0.0",
     "cypress": "^13.15.0",
     "dotenv-webpack": "^8.1.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.18.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-prettier": "^10.0.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-mdx": "^3.1.5",
     "eslint-plugin-css-modules": "^2.12.0",
     "eslint-plugin-cypress": "^4.0.0",
     "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.10.0",
     "eslint-plugin-jsdoc": "^50.4.3",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-mdx": "^3.1.5",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "4.6.2",
-    "eslint-plugin-storybook": "^0.10.1",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-storybook": "^0.11.2",
     "express": "^4.21.1",
     "express-static-gzip": "^2.1.8",
+    "globals": "^15.14.0",
     "html-webpack-plugin": "^5.6.3",
     "husky": "^9.1.6",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "knip": "^5.33.3",
+    "knip": "^5.33.2",
     "lint-staged": "^15.2.10",
     "mini-css-extract-plugin": "^2.9.1",
     "postcss-loader": "^8.1.1",
     "postcss-preset-env": "^10.0.7",
     "prettier": "^3.3.3",
     "process": "^0.11.10",
-    "release-it": "^17.10.0",
+    "release-it": "^18.1.1",
     "rimraf": "^6.0.1",
     "sass": "^1.80.3",
     "sass-loader": "^16.0.2",
@@ -206,21 +216,25 @@
     "webpack": "^5.95.0",
     "webpack-assets-manifest": "^5.2.1",
     "webpack-bundle-analyzer": "^4.10.2",
-    "webpack-cli": "^5.1.4",
+    "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.1.0",
     "webpack-merge": "^6.0.1",
     "workbox-webpack-plugin": "^7.1.0"
   },
   "dependencies": {
-    "@arpitmalik832/react-js-rollup-monorepo-library": "^0.3.10-alpha.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.27.0",
+    "@arpitmalik832/react-js-rollup-library": "0.0.1-alpha.5",
+    "@reduxjs/toolkit": "^2.5.0",
+    "axios": "^1.7.9",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-redux": "^9.2.0",
+    "react-router": "^7.1.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^3.1.0",
     "workbox-cacheable-response": "^7.1.0",
     "workbox-precaching": "^7.1.0",
     "workbox-routing": "^7.1.0",
     "workbox-strategies": "^7.1.0"
-  }
+  },
+  "packageManager": "pnpm@10.0.0-beta.1"
 }
