@@ -6,7 +6,7 @@
 
 /// JSON fixture file can be loaded directly using
 // the built-in JavaScript bundler
-const requiredExample = require('../../fixtures/example.json');
+import requiredExample from '../../fixtures/example.json';
 
 context('Files', () => {
   beforeEach(() => {
@@ -43,9 +43,7 @@ context('Files', () => {
     // we are inside the "function () { ... }"
     // callback and can use test context object "this"
     // "this.example" was loaded in "beforeEach" function callback
-    expect(this.example, 'fixture in the test context').to.deep.equal(
-      requiredExample,
-    );
+    cy.wrap(this.example).should('deep.equal', requiredExample);
 
     // or use "cy.wrap" and "should('deep.equal', ...)" assertion
     cy.wrap(this.example).should('deep.equal', requiredExample);
@@ -57,7 +55,7 @@ context('Files', () => {
     // You can read a file and yield its contents
     // The filePath is relative to your project's root.
     cy.readFile(Cypress.config('configFile')).then(config => {
-      expect(config).to.be.an('string');
+      cy.wrap(config).should('be.an', 'string');
     });
   });
 
@@ -72,9 +70,8 @@ context('Files', () => {
       cy.writeFile('cypress/fixtures/users.json', response.body);
     });
 
-    cy.fixture('users').should(users => {
-      // eslint-disable-next-line no-unused-expressions
-      expect(users[0].name).to.exist;
+    cy.fixture('users').then(users => {
+      cy.wrap(users[0].name).should('exist');
     });
 
     // JavaScript arrays and objects are stringified
@@ -85,8 +82,8 @@ context('Files', () => {
       email: 'jane@example.com',
     });
 
-    cy.fixture('profile').should(profile => {
-      expect(profile.name).to.eq('Jane');
+    cy.fixture('profile').then(profile => {
+      cy.wrap(profile.name).should('eq', 'Jane');
     });
   });
 });
