@@ -19,7 +19,7 @@ export default {
     'storybook-addon-render-modes',
   ],
   framework: '@storybook/react-webpack5',
-  webpackFinal: async (config, { configType }) => {
+  webpackFinal: async config => {
     if (!process.env.STORY_ENV) {
       throw new Error(ERR_NO_STORY_ENV_FLAG);
     }
@@ -69,6 +69,7 @@ export default {
     });
 
     // adding code splitting
+    // eslint-disable-next-line no-param-reassign
     config.optimization = {
       ...config.optimization,
       minimize: isRelease || isBeta,
