@@ -42,9 +42,9 @@ context('Connectors', () => {
     const arr = ['foo', 'bar', 'baz'];
 
     cy.wrap(arr).spread((foo, bar, baz) => {
-      expect(foo).to.eq('foo');
-      expect(bar).to.eq('bar');
-      expect(baz).to.eq('baz');
+      cy.wrap(foo).should('eq', 'foo');
+      cy.wrap(bar).should('eq', 'bar');
+      cy.wrap(baz).should('eq', 'baz');
     });
   });
 
@@ -52,41 +52,41 @@ context('Connectors', () => {
     it('invokes a callback function with the current subject', () => {
       // https://on.cypress.io/then
       cy.get('.connectors-list > li').then($lis => {
-        expect($lis, '3 items').to.have.length(3);
-        expect($lis.eq(0), 'first item').to.contain('Walk the dog');
-        expect($lis.eq(1), 'second item').to.contain('Feed the cat');
-        expect($lis.eq(2), 'third item').to.contain('Write JavaScript');
+        cy.wrap($lis).should('have.length', 3);
+        cy.wrap($lis.eq(0)).should('contain', 'Walk the dog');
+        cy.wrap($lis.eq(1)).should('contain', 'Feed the cat');
+        cy.wrap($lis.eq(2)).should('contain', 'Write JavaScript');
       });
     });
 
     it('yields the returned value to the next command', () => {
       cy.wrap(1)
         .then(num => {
-          expect(num).to.equal(1);
+          cy.wrap(num).should('eq', 1);
 
-          return 2;
+          return cy.wrap(2);
         })
         .then(num => {
-          expect(num).to.equal(2);
+          cy.wrap(num).should('eq', 2);
         });
     });
 
     it('yields the original subject without return', () => {
       cy.wrap(1)
         .then(num => {
-          expect(num).to.equal(1);
+          cy.wrap(num).should('eq', 1);
           // note that nothing is returned from this callback
         })
         .then(num => {
           // this callback receives the original unchanged value 1
-          expect(num).to.equal(1);
+          cy.wrap(num).should('eq', 1);
         });
     });
 
     it('yields the value yielded by the last Cypress command inside', () => {
       cy.wrap(1)
         .then(num => {
-          expect(num).to.equal(1);
+          cy.wrap(num).should('eq', 1);
           // note how we run a Cypress command
           // the result yielded by this Cypress command
           // will be passed to the second ".then"
@@ -94,7 +94,7 @@ context('Connectors', () => {
         })
         .then(num => {
           // this callback receives the value yielded by "cy.wrap(2)"
-          expect(num).to.equal(2);
+          cy.wrap(num).should('eq', 2);
         });
     });
   });
