@@ -2,8 +2,6 @@
  * This file contains examples of Cypress APIs in action.
  * @file The file is saved as `cypress/e2e/2-advanced-examples/cypress_api.cy.js`.
  */
-/* eslint-disable no-unused-expressions */
-/* eslint-disable no-param-reassign */
 /* eslint-disable no-console */
 /// <reference types="cypress" />
 
@@ -25,11 +23,8 @@ context('Cypress APIs', () => {
           // the previous subject is automatically received
           // and the commands arguments are shifted
 
-          // allow us to change the console method used
-          method = method || 'log';
-
           // log the subject to the console
-          console[method]('The subject is', subject);
+          console[method || 'log']('The subject is', subject);
 
           // whatever we return becomes the new subject
           // we don't want to change the subject so
@@ -72,7 +67,7 @@ context('Cypress APIs', () => {
 
     it('Get CPU architecture name of underlying OS', () => {
       // https://on.cypress.io/arch
-      expect(Cypress.arch).to.exist;
+      cy.wrap(Cypress.arch).should('exist');
     });
   });
 
@@ -85,22 +80,26 @@ context('Cypress APIs', () => {
       // https://on.cypress.io/config
       const myConfig = Cypress.config();
 
-      expect(myConfig).to.have.property('animationDistanceThreshold', 5);
-      expect(myConfig).to.have.property('baseUrl', null);
-      expect(myConfig).to.have.property('defaultCommandTimeout', 4000);
-      expect(myConfig).to.have.property('requestTimeout', 5000);
-      expect(myConfig).to.have.property('responseTimeout', 30000);
-      expect(myConfig).to.have.property('viewportHeight', 660);
-      expect(myConfig).to.have.property('viewportWidth', 1000);
-      expect(myConfig).to.have.property('pageLoadTimeout', 60000);
-      expect(myConfig).to.have.property('waitForAnimations', true);
+      cy.wrap(myConfig).should(
+        'have.property',
+        'animationDistanceThreshold',
+        5,
+      );
+      cy.wrap(myConfig).should('have.property', 'baseUrl', null);
+      cy.wrap(myConfig).should('have.property', 'defaultCommandTimeout', 4000);
+      cy.wrap(myConfig).should('have.property', 'requestTimeout', 5000);
+      cy.wrap(myConfig).should('have.property', 'responseTimeout', 30000);
+      cy.wrap(myConfig).should('have.property', 'viewportHeight', 660);
+      cy.wrap(myConfig).should('have.property', 'viewportWidth', 1000);
+      cy.wrap(myConfig).should('have.property', 'pageLoadTimeout', 60000);
+      cy.wrap(myConfig).should('have.property', 'waitForAnimations', true);
 
-      expect(Cypress.config('pageLoadTimeout')).to.eq(60000);
+      cy.wrap(Cypress.config('pageLoadTimeout')).should('eq', 60000);
 
       // this will change the config for the rest of your tests!
       Cypress.config('pageLoadTimeout', 20000);
 
-      expect(Cypress.config('pageLoadTimeout')).to.eq(20000);
+      cy.wrap(Cypress.config('pageLoadTimeout')).should('eq', 20000);
 
       Cypress.config('pageLoadTimeout', 60000);
     });
@@ -117,8 +116,8 @@ context('Cypress APIs', () => {
       const visibleP = Cypress.$('.dom-p p.visible').get(0);
 
       // our first paragraph has css class 'hidden'
-      expect(Cypress.dom.isHidden(hiddenP)).to.be.true;
-      expect(Cypress.dom.isHidden(visibleP)).to.be.false;
+      cy.wrap(Cypress.dom.isHidden(hiddenP)).should('be.true');
+      cy.wrap(Cypress.dom.isHidden(visibleP)).should('be.false');
     });
   });
 
@@ -139,15 +138,23 @@ context('Cypress APIs', () => {
       });
 
       // get environment variable
-      expect(Cypress.env('host')).to.eq('veronica.dev.local');
+      cy.wrap(Cypress.env('host')).should('eq', 'veronica.dev.local');
 
       // set environment variable
       Cypress.env('api_server', 'http://localhost:8888/v2/');
-      expect(Cypress.env('api_server')).to.eq('http://localhost:8888/v2/');
+      cy.wrap(Cypress.env('api_server')).should(
+        'eq',
+        'http://localhost:8888/v2/',
+      );
 
       // get all environment variable
-      expect(Cypress.env()).to.have.property('host', 'veronica.dev.local');
-      expect(Cypress.env()).to.have.property(
+      cy.wrap(Cypress.env()).should(
+        'have.property',
+        'host',
+        'veronica.dev.local',
+      );
+      cy.wrap(Cypress.env()).should(
+        'have.property',
         'api_server',
         'http://localhost:8888/v2/',
       );
@@ -171,7 +178,7 @@ context('Cypress APIs', () => {
 
     it('Get underlying OS name', () => {
       // https://on.cypress.io/platform
-      expect(Cypress.platform).to.be.exist;
+      cy.wrap(Cypress.platform).should('be.exist');
     });
   });
 
@@ -182,7 +189,7 @@ context('Cypress APIs', () => {
 
     it('Get current version of Cypress being run', () => {
       // https://on.cypress.io/version
-      expect(Cypress.version).to.be.exist;
+      cy.wrap(Cypress.version).should('be.exist');
     });
   });
 
