@@ -2,7 +2,6 @@
  * This test file demonstrates how to use Cypress commands to interact with the location.
  * @file The file is saved as `cypress/e2e/2-advanced-examples/location.cy.js`.
  */
-/* eslint-disable no-unused-expressions */
 /// <reference types="cypress" />
 
 context('Location', () => {
@@ -17,18 +16,19 @@ context('Location', () => {
 
   it('cy.location() - get window.location', () => {
     // https://on.cypress.io/location
-    cy.location().should(location => {
-      expect(location.hash).to.be.empty;
-      expect(location.href).to.eq(
+    cy.location().then(location => {
+      cy.wrap(location.hash).should('be.empty');
+      cy.wrap(location.href).should(
+        'eq',
         'https://example.cypress.io/commands/location',
       );
-      expect(location.host).to.eq('example.cypress.io');
-      expect(location.hostname).to.eq('example.cypress.io');
-      expect(location.origin).to.eq('https://example.cypress.io');
-      expect(location.pathname).to.eq('/commands/location');
-      expect(location.port).to.eq('');
-      expect(location.protocol).to.eq('https:');
-      expect(location.search).to.be.empty;
+      cy.wrap(location.host).should('eq', 'example.cypress.io');
+      cy.wrap(location.hostname).should('eq', 'example.cypress.io');
+      cy.wrap(location.origin).should('eq', 'https://example.cypress.io');
+      cy.wrap(location.pathname).should('eq', '/commands/location');
+      cy.wrap(location.port).should('eq', '');
+      cy.wrap(location.protocol).should('eq', 'https:');
+      cy.wrap(location.search).should('be.empty');
     });
   });
 
