@@ -2,7 +2,6 @@
  * This file contains examples of navigation commands in Cypress.
  * @file The file is saved as `cypress/e2e/2-advanced-examples/navigation.cy.js`.
  */
-/* eslint-disable no-unused-expressions */
 /// <reference types="cypress" />
 
 context('Navigation', () => {
@@ -50,11 +49,11 @@ context('Navigation', () => {
       timeout: 50000, // increase total time for the visit to resolve
       onBeforeLoad(contentWindow) {
         // contentWindow is the remote page's window object
-        expect(typeof contentWindow === 'object').to.be.true;
+        assert.equal(typeof contentWindow, 'object');
       },
       onLoad(contentWindow) {
         // contentWindow is the remote page's window object
-        expect(typeof contentWindow === 'object').to.be.true;
+        assert.equal(typeof contentWindow, 'object');
       },
     });
   });
