@@ -3,7 +3,6 @@
  * @file The file is saved as `cypress/e2e/2-advanced-examples/spies_stubs_clocks.cy.js`.
  */
 /* eslint-disable no-console */
-/* eslint-disable no-unused-expressions */
 /// <reference types="cypress" />
 
 context('Spies, Stubs, and Clock', () => {
@@ -12,14 +11,16 @@ context('Spies, Stubs, and Clock', () => {
     cy.visit('https://example.cypress.io/commands/spies-stubs-clocks');
 
     const obj = {
-      foo() {},
+      foo() {
+        return undefined;
+      },
     };
 
     const spy = cy.spy(obj, 'foo').as('anyArgs');
 
     obj.foo();
 
-    expect(spy).to.be.called;
+    cy.wrap(spy).should('be.called');
   });
 
   it('cy.spy() retries until assertions pass', () => {
@@ -72,7 +73,7 @@ context('Spies, Stubs, and Clock', () => {
 
     obj.foo('foo', 'bar');
 
-    expect(stub).to.be.called;
+    cy.wrap(stub).should('be.called');
   });
 
   it('cy.clock() - control time in the browser', () => {
@@ -122,18 +123,25 @@ context('Spies, Stubs, and Clock', () => {
     };
 
     cy.stub(greeter, 'greet')
-      .callThrough() // if you want non-matched calls to call the real method
+      .callThrough()
       .withArgs(Cypress.sinon.match.string)
       .returns('Hi')
       .withArgs(Cypress.sinon.match.number)
       .throws(new Error('Invalid name'));
 
-    expect(greeter.greet('World')).to.equal('Hi');
-    expect(() => greeter.greet(42)).to.throw('Invalid name');
-    expect(greeter.greet).to.have.been.calledTwice;
+    // Test string argument
+    cy.wrap(greeter.greet('World')).should('equal', 'Hi');
 
-    // non-matched calls goes the actual method
-    expect(greeter.greet()).to.equal('Hello, undefined!');
+    // Test number argument
+    try {
+      greeter.greet(42);
+    } catch (err) {
+      cy.wrap(err.message).should('equal', 'Invalid name');
+    }
+
+    // Test unmatched call
+    const defaultResult = greeter.greet();
+    cy.wrap(defaultResult).should('equal', 'Hello, undefined!');
   });
 
   it('matches call arguments using Sinon matchers', () => {
@@ -155,28 +163,30 @@ context('Spies, Stubs, and Clock', () => {
 
     const spy = cy.spy(calculator, 'add').as('add');
 
-    expect(calculator.add(2, 3)).to.equal(5);
+    cy.wrap(calculator.add(2, 3)).should('equal', 5);
 
     // if we want to assert the exact values used during the call
-    expect(spy).to.be.calledWith(2, 3);
+    cy.wrap(spy).should('be.calledWith', 2, 3);
 
     // let's confirm "add" method was called with two numbers
-    expect(spy).to.be.calledWith(
+    cy.wrap(spy).should(
+      'be.calledWith',
       Cypress.sinon.match.number,
       Cypress.sinon.match.number,
     );
 
     // alternatively, provide the value to match
-    expect(spy).to.be.calledWith(
+    cy.wrap(spy).should(
+      'be.calledWith',
       Cypress.sinon.match(2),
       Cypress.sinon.match(3),
     );
 
     // match any value
-    expect(spy).to.be.calledWith(Cypress.sinon.match.any, 3);
+    cy.wrap(spy).should('be.calledWith', Cypress.sinon.match.any, 3);
 
     // match any value from a list
-    expect(spy).to.be.calledWith(Cypress.sinon.match.in([1, 2, 3]), 3);
+    cy.wrap(spy).should('be.calledWith', Cypress.sinon.match.in([1, 2, 3]), 3);
 
     /**
      * Returns true if the given number is even.
@@ -191,7 +201,11 @@ context('Spies, Stubs, and Clock', () => {
     // expect the value to pass a custom predicate function
     // the second argument to "sinon.match(predicate, message)" is
     // shown if the predicate does not pass and assertion fails
-    expect(spy).to.be.calledWith(Cypress.sinon.match(isEven, 'isEven'), 3);
+    cy.wrap(spy).should(
+      'be.calledWith',
+      Cypress.sinon.match(isEven, 'isEven'),
+      3,
+    );
 
     /**
      * Returns a function that checks if a given number is larger than the limit.
@@ -216,14 +230,16 @@ context('Spies, Stubs, and Clock', () => {
     const isLessThan = limit => x => x < limit;
 
     // you can combine several matchers using "and", "or"
-    expect(spy).to.be.calledWith(
+    cy.wrap(spy).should(
+      'be.calledWith',
       Cypress.sinon.match.number,
       Cypress.sinon
         .match(isGreaterThan(2), '> 2')
         .and(Cypress.sinon.match(isLessThan(4), '< 4')),
     );
 
-    expect(spy).to.be.calledWith(
+    cy.wrap(spy).should(
+      'be.calledWith',
       Cypress.sinon.match.number,
       Cypress.sinon
         .match(isGreaterThan(200), '> 200')
