@@ -1,8 +1,4 @@
-/**
- * Button storybook component.
- * @file The file is saved as `Button/index.stories.jsx`.
- */
-import { ReduxProvider } from '@arpitmalik832/react-js-rollup-monorepo-library';
+import { ReduxProvider } from '@arpitmalik832/react-js-rollup-library';
 
 import Button from './index';
 import store from '../../../redux/store/store';
