@@ -7,9 +7,9 @@ import '@testing-library/jest-dom';
 
 import PageWrapper from '../PageWrapper';
 
-jest.mock('react-router-dom', () => ({
+jest.mock('react-router', () => ({
   __esModule: true,
-  ...jest.requireActual('react-router-dom'),
+  ...jest.requireActual('react-router'),
   useLocation: () => ({
     pathname: '/test',
   }),
