@@ -7,10 +7,14 @@ import '@testing-library/jest-dom';
 
 import hook from '../useAppMount';
 
-jest.mock('@arpitmalik832/react-js-rollup-monorepo-library', () => ({
+jest.mock('@arpitmalik832/react-js-rollup-library', () => ({
   __esModule: true,
   useTheme: jest.fn(),
-  useInitAxios: jest.fn(),
+}));
+
+jest.mock('../useInitAxios', () => ({
+  __esModule: true,
+  default: jest.fn(),
 }));
 
 describe('useAppMount unit tests', () => {
