@@ -2,19 +2,36 @@
  * This is the entry point of the application.
  * @file This file is saved as `index.js`.
  */
-import { SWRegistration } from '@arpitmalik832/react-js-rollup-monorepo-library';
+import {
+  errorLog,
+  SWRegistration,
+} from '@arpitmalik832/react-js-rollup-library';
 
 import { ENVS } from '../build_utils/config/index.mjs';
 
 // eslint-disable-next-line import/extensions
-import('./bootstrap.jsx').then(({ mount }) =>
-  mount(document.getElementById('app')),
-);
+import('./bootstrap.jsx')
+  .then(({ mount }) => {
+    const appElement = document.getElementById('app');
+    if (appElement) {
+      mount(appElement);
+    } else {
+      errorLog('App element not found');
+    }
+  })
+  .catch(err => {
+    errorLog('Facing err while importing bootstrap file: ', err);
+  });
 
 SWRegistration.register();
 
 if (process.env.APP_ENV !== ENVS.PROD) {
-  import('@arpitmalik832/react-js-rollup-monorepo-library').then(
-    ({ reportWebVitals: func }) => func(),
-  );
+  import('@arpitmalik832/react-js-rollup-library')
+    .then(({ reportWebVitals: func }) => func())
+    .catch(err => {
+      errorLog(
+        'Facing issue while using reportWebVitals from external library',
+        err,
+      );
+    });
 }
