@@ -5,9 +5,14 @@
 import {
   useBackPress,
   Button,
-} from '@arpitmalik832/react-js-rollup-monorepo-library';
+  log,
+} from '@arpitmalik832/react-js-rollup-library';
+import { useSelector } from 'react-redux';
+import { useEffect } from 'react';
 
 import ButtonV2 from '../../components/atoms/Button';
+import { ReactComponent as ReactIcon } from '../../assets/icons/react.svg';
+import { useFetchDataQuery } from '../../redux/queries/sampleQuery';
 
 /**
  * Home component renders the home page with buttons.
@@ -16,13 +21,23 @@ import ButtonV2 from '../../components/atoms/Button';
  * <Home />
  */
 function Home() {
+  const apis = useSelector(state => state.apis);
+
   useBackPress();
+  const { data, isLoading, isError } = useFetchDataQuery(
+    apis[0]?.axiosInstance,
+  );
+
+  useEffect(() => {
+    log({ isLoading, data, isError });
+  }, [isLoading, data, isError]);
 
   return (
     <div>
       Home
       <Button />
       <ButtonV2 />
+      <ReactIcon />
     </div>
   );
 }
