@@ -4,27 +4,30 @@
  */
 import { thunk } from 'redux-thunk';
 import { logger } from 'redux-logger';
-import {
-  slices,
-  queries,
-  configureStore,
-} from '@arpitmalik832/react-js-rollup-monorepo-library';
+import { slices } from '@arpitmalik832/react-js-rollup-library';
+import { configureStore } from '@reduxjs/toolkit';
+// eslint-disable-next-line import/extensions
+import { sampleQuery } from '../queries/sampleQuery.js';
 
-export default configureStore({
+const store = configureStore({
   reducer: {
     app: slices.appSlice.reducer,
     apis: slices.apisSlice.reducer,
     navigation: slices.navigationSlice.reducer,
-    sampleQuery: queries.sampleQuery.reducer,
+    sampleQuery: sampleQuery.reducer,
   },
   middleware: getDefault =>
     getDefault({
       serializableCheck: {
         ignoredActions: [
-          'apis/updateApi1AxiosInstance',
+          'apis/addNewApiData',
           'navigation/pushStack',
+          'sampleQuery/executeQuery/rejected',
+          'sampleQuery/executeQuery/fulfilled',
         ],
         ignoredPaths: ['apis', 'sampleQuery', 'navigation'],
       },
-    }).concat(queries.sampleQuery.middleware, thunk, logger),
+    }).concat(sampleQuery.middleware, thunk, logger),
 });
+
+export default store;
