@@ -3,8 +3,8 @@
  * @file This file is saved as `routes/index.jsx`.
  */
 import { lazy } from 'react';
-import { createBrowserRouter } from 'react-router-dom';
-import { ComponentWithSuspense } from '@arpitmalik832/react-js-rollup-monorepo-library';
+import { createBrowserRouter } from 'react-router';
+import { ComponentWithSuspense } from '@arpitmalik832/react-js-rollup-library';
 
 import PageWrapper from '../components/organisms/PageWrapper';
 import routes from './routes';
